refactoring + optimisation

+++ src/main/BiAlgorithm.java
@@ -3,5 +3,5 @@ package main;
 import java.util.List;
 
 public interface BiAlgorithm {
-public List<List<Integer>> execute(int[][] G, float d);
+public List<List<Integer>> execute(int[][] G, double d);
 }
